knowledgeGraph/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blis==1.0.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.8
decorator==5.1.1
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
gunicorn==23.0.0
httplib2==0.22.0
huggingface-hub==0.26.0
idna==3.10
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.10
neo4j==5.25.0
nest-asyncio==1.6.0
networkx==3.4.1
nltk==3.9.1
numpy==2.0.2
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
preshed==3.0.9
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.3
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
rank-bm25==0.2.2
regex==2024.9.11
requests==2.32.3
rich==13.9.2
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
stack-data==0.6.3
sympy==1.13.1
thinc==8.3.2
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
underthesea==6.8.4
underthesea_core==1.0.4
uritemplate==4.1.1
urllib3==2.2.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
whitenoise==6.8.1
wrapt==1.16.0
yarl==1.16.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
